+++ blocklets/core/api/src/libs/auth.ts
@@ -2,7 +2,7 @@ import path from 'path';
 
 import AuthStorage from '@arcblock/did-connect-storage-nedb';
 import env from '@blocklet/sdk/lib/env';
-import AuthService from '@blocklet/sdk/lib/service/auth';
+import AuthService from '@blocklet/sdk/lib/service/blocklet';
 import getWallet from '@blocklet/sdk/lib/wallet';
 import WalletAuthenticator from '@blocklet/sdk/lib/wallet-authenticator';
 import WalletHandler from '@blocklet/sdk/lib/wallet-handler';