+++ blocklets/core/api/src/libs/notifications/templates/index.ts
@@ -1,2 +1,3 @@
 export * from './base';
 export * from './credit-granted';
+export * from './credential';