+++ blocklets/core/src/pages/config/ai-config/ai-model-rates/index.tsx
@@ -169,13 +169,18 @@ export default function AIModelRates() {
 
   useSubscription(
     'model.status.updated',
-    ({ provider, model, available }: { provider: string; model: string; available: boolean }) => {
+    ({ provider, model, available, error }: { provider: string; model: string; available: boolean; error: any }) => {
       mutate((r: any) => {
         const data = r.list || [];
+
         data.forEach((item: any) => {
-          if (item.provider?.name === provider && item.model === model && item.status) {
+          if (item.provider?.name === provider && item.model === model) {
             item.loading = false;
-            item.status.available = available;
+
+            if (item.status) {
+              item.status.available = available;
+              item.status.error = error;
+            }
           }
         });