+++ blocklets/core/src/pages/pricing/index.tsx
@@ -137,16 +137,23 @@ export default function PricingPage() {
 
   useSubscription(
     'model.status.updated',
-    ({ provider, model, available }: { provider: string; model: string; available: boolean }) => {
+    ({ provider, model, available, error }: { provider: string; model: string; available: boolean; error: any }) => {
       mutate((r: any) => {
-        r.forEach((item: any) => {
+        return r.map((item: any) => {
           if (item.provider === provider && item.model === model && item.status) {
-            item.loading = false;
-            item.status.available = available;
+            return {
+              ...item,
+              loading: false,
+              status: {
+                ...item.status,
+                available,
+                error,
+              },
+            };
           }
-        });
 
-        return r;
+          return item;
+        });
       });
     },
     []