+++ blocklets/core/CHANGELOG.md
@@ -1,6 +1,10 @@
+## 0.4.5 (2025-9-2)
+
+- chore: check credential before add credential
+
 ## 0.4.4 (2025-9-2)
 
-- chore: merge model for chat
+- chore: merge model for chat api
 
 ## 0.4.3 (2025-9-1)